added tcp connection keep alive using WAMP raw socket ping-pong.
- PING_TIMEOUT = 30 secs
- TIMEOUT = PING_TIMEOUT * 2
- PING_MAX_ATTEMPTS = 2

A timeout triggers a ping. If pong is not received in PING_TIMEOUT another ping is sent until we reach PING_MAX_ATTEMPTS. If the pong is not received in PING_TIMEOUT and we have reached PING_MAX_ATTEMPTS the connection is closed. Every time we receive a PING_TIMEOUT the gen_server resets its TIMEOUT.