Fixes an bug that occus when subscriber (session) is subscribing to the same topic|options multiple times
A session subscribing multiple times on the same topic|options is a valid use case as it enables the client to implement multiple dispatch of events.

The issue was that subscribe request were handled asynchronoulsy which can cause two or more subscribe requests to create a new subscription concurrently as they all check for a matchin subscription and all getting a not found thus creating a new subscription.

Also fixed a bug in the pattern matching of the returned registry entry.

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>