Added support for internal callback peers
This is an option for Bondy to internally define a callee callback module (conforming to bondy_wamp_api behaviour) as an RPC procedure peer. This is to allow making calls to modules in other nodes.

The new API also adds the capability to define internal process callees (but not completely implemented or used yet).

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>