create new profile to be able to build releases were the nodename and cookie are hardcoded
This is to avoid having the /bondy/releases to be read-only which is a desirable property on embedded systems. We need to do this since rebar3/relx do not allow us to specify the location of the vm.args and we cannot set name and cookie on the included file generated by scuttler/cuttlefish.

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>