complete use of rebase to reduce variations on statically defined realm configuration
This is still not perfect as signing and encryption keys are not required, so the realm will generate default values. On a cluster deployment all nodes will generate differing valus for the same statically defined realm and thus this will trigger the replication of the new values. The last deployment will override all previous ones. This will mean that if using tickets, the previous tickets will be invalidated (actually depending on ticket configuration) as the signing keys for the realm have changed.  To avoid this the user needs to provide the keys e.g. some secret management tool injecting the keys as ENV VARS to be picked up by the configuration file.

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>